python/reflection/requirements.txt
grpcio==1.47.0
grpcio-reflection==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.1
six==1.16.0
